Added more change set functions
- The change set function for determining such a structure had to be
  reworked in order for the apply function to work properly. The
  determine function must take into consideration the appropriate
  indices for each separate step, such that the reconstruction works as
  intended.
- Fixed the Deck's iterator missing virtual functions
- The BooleanBitField API has been improved a bit
- The entity manager can now delete multiple archetypes from an iterator
  instead

+++ ECSEngine/src/ECSEngine/ECS/EntityManagerChangeSet.h
@@ -122,7 +122,7 @@ namespace ECSEngine {
 		// The order in which these are applied is important. Firstly, the necessary archetypes are destroyed,
 		// Then the new archetypes are created and at last the moved archetypes are handled. It is critical that
 		// These operations are done exactly in this order, otherwise the consistency is not achieved
-		DeckPowerOfTwo<unsigned int> destroyed_archetypes;
+		DeckPowerOfTwo<unsigned short> destroyed_archetypes;
 		DeckPowerOfTwo<NewArchetype> new_archetypes;
 		DeckPowerOfTwo<MovedElementIndex<unsigned short>> moved_archetypes;
